Adds controllers and routes * Validations * Middleware for error handling
