+++ packages/rest-api/src/constants/index.ts
@@ -1,0 +1,5 @@
+export const VALID_BRIDGE_MODULES = [
+  'SynapseBridge',
+  'SynapseCCTP',
+  'SynapseRFQ',
+]
