+++ packages/rest-api/src/controllers/swapController.ts
@@ -1,0 +1,30 @@
+import { validationResult } from 'express-validator'
+import { formatUnits, parseUnits } from '@ethersproject/units'
+
+import { Synapse } from '../services/synapseService'
+
+export const swapController = async (req, res) => {
+  const errors = validationResult(req)
+  if (!errors.isEmpty()) {
+    return res.status(400).json({ errors: errors.array() })
+  }
+  try {
+    const { chain, amount } = req.query
+    const fromTokenInfo = res.locals.tokenInfo.fromToken
+    const toTokenInfo = res.locals.tokenInfo.toToken
+
+    const amountInWei = parseUnits(amount.toString(), fromTokenInfo.decimals)
+    const quote = await Synapse.swapQuote(
+      Number(chain),
+      fromTokenInfo.address,
+      toTokenInfo.address,
+      amountInWei
+    )
+    res.json({
+      maxAmountOut: formatUnits(quote.maxAmountOut, toTokenInfo.decimals),
+      ...quote,
+    })
+  } catch (err) {
+    res.status(500).json({ error: 'Server error' })
+  }
+}
