+++ packages/rest-api/src/controllers/tokensListController.ts
@@ -1,0 +1,5 @@
+import * as tokenList from '../constants/bridgeable'
+
+export const tokenListController = async (_req, res) => {
+  res.json(tokenList)
+}
