+++ packages/rest-api/src/routes/indexRoute.ts
@@ -1,0 +1,9 @@
+import express from 'express'
+
+import { indexController } from '../controllers/indexController'
+
+const router = express.Router()
+
+router.get('/', indexController)
+
+export default router
