+++ packages/rest-api/src/routes/tokenListRoute.ts
@@ -1,0 +1,9 @@
+import express from 'express'
+
+import { tokenListController } from '../controllers/tokensListController'
+
+const router = express.Router()
+
+router.get('/', tokenListController)
+
+export default router
